Add map filters + refactoring & style fixes (#432)
