+++ src/ui/views/common/BaseModal.js
@@ -12,10 +12,10 @@ const styles = {
     display: 'flex',
     flexDirection: 'column',
     position: 'absolute',
+    height: '700px',
     width: '700px',
-    maxWidth: '60%',
-    height: '800px',
     maxHeight: '60%',
+    maxWidth: '60%',
     p: 2,
     color: 'text.primary',
     bgcolor: 'background.paper',
