+++ src/ui/views/common/FilterModal.js
@@ -15,50 +15,22 @@ import {
   ButtonGroup,
   TextField,
 } from '@mui/material'
-import {
-  DataGrid,
-  GridActionsCellItem,
-  GridRowModes,
-  GRID_CHECKBOX_SELECTION_COL_DEF,
-  useGridApiRef,
-  GridRowEditStopReasons,
-} from '@mui/x-data-grid'
 import * as R from 'ramda'
-import { Fragment, useCallback, useEffect, useMemo, useState } from 'react'
-import {
-  MdAddCircleOutline,
-  MdArrowForwardIos,
-  MdCheck,
-  MdDelete,
-  MdEdit,
-  MdRestore,
-  MdSave,
-} from 'react-icons/md'
+import { Fragment, useCallback, useMemo, useState } from 'react'
+import { MdArrowForwardIos, MdCheck, MdRestore } from 'react-icons/md'
 import { useSelector } from 'react-redux'
 
 import { DataGridModal } from './BaseModal'
+import GridFilter from './GridFilter'
 
 import {
   selectGroupedOutputNames,
-  selectNumberFormat,
   selectStatGroupings,
 } from '../../../data/selectors'
-import OverflowText from '../../compound/OverflowText'
 
-import {
-  NumberFormat,
-  getLabelFn,
-  getSubLabelFn,
-  mapIndexed,
-  renameKeys,
-} from '../../../utils'
+import { getLabelFn, getSubLabelFn } from '../../../utils'
 
 const styles = {
-  addBtn: {
-    justifyContent: 'start',
-    pl: 2,
-    py: 2,
-  },
   content: {
     display: 'flex',
     flexDirection: 'column',
@@ -68,17 +40,12 @@ const styles = {
     textAlign: 'center',
     overflow: 'auto',
   },
-  emptyContent: {
-    display: 'flex',
-    flexDirection: 'column',
-    alignItems: 'center',
-    justifyContent: 'center',
-    bgcolor: 'rgba(18, 18, 18, 0.38)',
-    height: '100%',
-  },
   accordRoot: {
     border: '1px solid',
     borderColor: 'divider',
+    '& .MuiAccordion-region': {
+      bgcolor: 'rgba(0, 0, 0, .125)',
+    },
     '&:not(:last-child)': {
       borderBottom: 0,
     },
@@ -87,8 +54,11 @@ const styles = {
     },
   },
   accordSummary: {
-    backgroundColor: 'rgba(255, 255, 255, .05)',
+    bgcolor: 'rgba(255, 255, 255, .05)',
     flexDirection: 'row-reverse',
+    '&.MuiAccordionSummary-root': {
+      mr: 0,
+    },
     '& .MuiAccordionSummary-expandIconWrapper.Mui-expanded': {
       transform: 'rotate(90deg)',
     },
@@ -106,369 +76,10 @@ const styles = {
     borderTop: '1px solid rgba(0, 0, 0, .125)',
   },
   leaves: {
-    // bgcolor: 'rgba(0, 0, 0, .125)',
-    // borderTop: '1px solid rgba(0, 0, 0, .125)',
     display: 'flex',
     flexWrap: 'wrap',
     ml: 4,
   },
-}
-
-const StatsFilter = ({ defaultFilters, onSave }) => {
-  const [filters, setFilters] = useState(defaultFilters)
-  const [idCount, setIdCount] = useState(0)
-  const [rows, setRows] = useState([])
-  const [initialRows, setInitialRows] = useState([])
-  const [rowModesModel, setRowModesModel] = useState({})
-  const [canSaveRow, setCanSaveRow] = useState({})
-
-  const apiRef = useGridApiRef()
-  const numberFormat = useSelector(selectNumberFormat)
-  const statNames = useSelector(selectGroupedOutputNames)
-
-  const isApiRefValid = !R.either(R.isNil, R.isEmpty)(apiRef.current)
-
-  const restoreFilters = useCallback(() => {
-    const filterCriteria = R.filter(
-      R.pipe(R.propOr('stat', 'format'), R.equals('stat'))
-    )(filters)
-    const initRows = mapIndexed(
-      R.pipe(
-        R.flip(R.assoc('id')),
-        // Set up `logic` for first row in case it is `undefined`
-        R.when(R.propEq(0, 'id'), R.assoc('logic', 'and')),
-        // Drop truthy `active` values
-        R.when(R.propEq(true, 'active'), R.dissoc('active')),
-        renameKeys({ option: 'relation', prop: 'source' })
-      )
-    )(filterCriteria)
-
-    setInitialRows(initRows)
-    setRows(initRows)
-    setIdCount(initRows.length)
-  }, [filters])
-
-  useEffect(() => {
-    restoreFilters()
-  }, [restoreFilters])
-
-  useEffect(() => {
-    if (!isApiRefValid) return
-
-    const newSelectedRowIds = R.pipe(
-      R.filter(R.propOr(true, 'active')),
-      R.pluck('id')
-    )(initialRows)
-    apiRef.current.setRowSelectionModel(newSelectedRowIds)
-  }, [apiRef, initialRows, isApiRefValid])
-
-  const handleAddRow = useCallback(() => {
-    setRows(
-      R.append({
-        isNew: true,
-        id: idCount,
-        source: '',
-        relation: '',
-        logic: 'and',
-      })
-    )
-    setRowModesModel(
-      R.assoc(idCount, {
-        mode: GridRowModes.Edit,
-        fieldToFocus: rows.length < 1 ? 'source' : 'logic',
-      })
-    )
-    apiRef.current.selectRow(idCount)
-    setIdCount(idCount + 1)
-  }, [apiRef, idCount, rows.length])
-
-  const deleteRow = useCallback((id) => {
-    setRows(R.reject(R.propEq(id, 'id')))
-  }, [])
-  const handleDeleteRow = useCallback(
-    (id) => () => {
-      deleteRow(id)
-    },
-    [deleteRow]
-  )
-
-  const handleClickDiscard = useCallback(
-    (id) => () => {
-      const row = apiRef.current.getRow(id)
-      setRowModesModel(
-        R.assoc(id, { mode: GridRowModes.View, ignoreModifications: true })
-      )
-      if (row.isNew) deleteRow(id)
-    },
-    [apiRef, deleteRow]
-  )
-  const handleClickEdit = useCallback(
-    (id) => () => {
-      setRowModesModel(R.assoc(id, { mode: GridRowModes.Edit }))
-    },
-    []
-  )
-  const handleClickSave = useCallback(
-    (id) => () => {
-      setRowModesModel(R.assoc(id, { mode: GridRowModes.View }))
-      setCanSaveRow(R.dissoc(id))
-    },
-    []
-  )
-
-  const processRowUpdate = (newRow) => {
-    const updatedRow = R.dissoc('isNew')(newRow)
-    setRows(R.map(R.when(R.propEq(newRow.id, 'id'), R.always(updatedRow))))
-    return updatedRow
-  }
-
-  const handleRowEditStop = ({ reason }, event) => {
-    if (reason === GridRowEditStopReasons.rowFocusOut) {
-      event.defaultMuiPrevented = true
-    }
-  }
-
-  const updateActiveStateForSelectedRow = useCallback(
-    (row) => {
-      const selectedRowIdsSet = new Set(apiRef.current.getSelectedRows().keys())
-      return R.ifElse(
-        (row) => selectedRowIdsSet.has(row.id),
-        R.dissoc('active'),
-        R.assoc('active', false)
-      )(row)
-    },
-    [apiRef]
-  )
-
-  const canDiscardOrSaveAll = useMemo(() => {
-    const unsavedChanges = !R.equals(rows)(initialRows)
-    const isEditing = R.pipe(
-      R.values,
-      R.any(R.propEq(GridRowModes.Edit, 'mode'))
-    )(rowModesModel)
-    return !isEditing && unsavedChanges
-  }, [initialRows, rowModesModel, rows])
-
-  const handleClickSaveAll = useCallback(() => {
-    const newFilters = R.map(
-      R.pipe(
-        updateActiveStateForSelectedRow,
-        R.dissoc('id'),
-        renameKeys({ relation: 'option', source: 'prop' })
-      )
-    )(rows)
-    setFilters(newFilters)
-    onSave(newFilters)
-  }, [onSave, rows, updateActiveStateForSelectedRow])
-
-  const handleRowSelectionCheckboxChange = useCallback(() => {
-    setRows(R.map(updateActiveStateForSelectedRow))
-  }, [updateActiveStateForSelectedRow])
-
-  const handleCellDoubleClick = (params, event) => {
-    event.defaultMuiPrevented = true
-  }
-
-  const preProcessEditCellProps = useCallback(
-    ({ id, props, hasChanged, otherFieldsProps }) => {
-      const isValueInvalid = R.propSatisfies(
-        R.either(R.isNil, R.isEmpty),
-        'value'
-      )
-      const newCellProps = R.pipe(
-        R.ifElse(isValueInvalid, R.assoc('error', true), R.dissoc('error'))
-      )(props)
-
-      if (hasChanged) {
-        const hasError =
-          newCellProps.error ||
-          R.pipe(
-            // We ignore the `logic` field in the first row
-            R.when(R.always(id === rows[0].id), R.dissoc('logic')),
-            R.values,
-            R.any(isValueInvalid)
-          )(otherFieldsProps)
-
-        setCanSaveRow(R.assoc(id, !hasError))
-      }
-      return newCellProps
-    },
-    [rows]
-  )
-
-  const columns = useMemo(
-    () => [
-      {
-        field: 'logic',
-        headerName: 'Logic',
-        align: 'center',
-        editable: true,
-        sortable: false,
-        type: 'singleSelect',
-        valueOptions: ['and'], // FIXME: ['or', 'and'],
-        getOptionLabel: (option) => option.toUpperCase(),
-        cellClassName: ({ id }) => (id > rows[0].id ? '' : 'hidden'),
-        preProcessEditCellProps,
-      },
-      GRID_CHECKBOX_SELECTION_COL_DEF,
-      {
-        field: 'source',
-        headerName: 'Statistic',
-        type: 'singleSelect',
-        editable: true,
-        sortable: false,
-        valueOptions: R.keys(statNames),
-        preProcessEditCellProps,
-        renderCell: ({ value }) => <OverflowText text={value} />,
-      },
-      {
-        field: 'relation',
-        headerName: 'Relation',
-        headerAlign: 'center',
-        align: 'center',
-        type: 'singleSelect',
-        editable: true,
-        sortable: false,
-        valueOptions: [
-          {
-            value: 'lt',
-            label: '<',
-          },
-          {
-            value: 'lte',
-            label: '<=',
-          },
-          {
-            value: 'eq',
-            label: '=',
-          },
-          {
-            value: 'gt',
-            label: '>',
-          },
-          {
-            value: 'gte',
-            label: '>=',
-          },
-        ],
-        preProcessEditCellProps,
-      },
-      {
-        field: 'value',
-        headerName: 'Value',
-        type: 'number',
-        editable: true,
-        sortable: false,
-        valueFormatter: ({ value }) =>
-          typeof value === 'number'
-            ? NumberFormat.format(value, numberFormat)
-            : null,
-        preProcessEditCellProps,
-        renderCell: ({ value }) => <OverflowText text={`${value}`} />,
-      },
-      {
-        field: 'actions',
-        headerName: 'Actions',
-        headerAlign: 'right',
-        align: 'right',
-        type: 'actions',
-        sortable: false,
-        getActions: ({ id }) =>
-          R.path([id, 'mode'])(rowModesModel) === GridRowModes.Edit
-            ? [
-                <GridActionsCellItem
-                  disabled={!canSaveRow[id]}
-                  icon={<MdSave size="20px" />}
-                  label="Save"
-                  onClick={handleClickSave(id)}
-                />,
-                <GridActionsCellItem
-                  icon={<MdRestore size="20px" />}
-                  label="Discard"
-                  onClick={handleClickDiscard(id)}
-                />,
-              ]
-            : [
-                <GridActionsCellItem
-                  icon={<MdEdit size="20px" />}
-                  label="Edit"
-                  onClick={handleClickEdit(id)}
-                />,
-                <GridActionsCellItem
-                  icon={<MdDelete size="20px" />}
-                  label="Delete"
-                  onClick={handleDeleteRow(id)}
-                />,
-              ],
-        preProcessEditCellProps,
-      },
-    ],
-    [
-      handleClickDiscard,
-      handleClickEdit,
-      handleClickSave,
-      handleDeleteRow,
-      canSaveRow,
-      numberFormat,
-      preProcessEditCellProps,
-      rowModesModel,
-      rows,
-      statNames,
-    ]
-  )
-
-  return (
-    <>
-      <Paper sx={styles.content}>
-        <DataGrid
-          {...{ apiRef, columns, rows, rowModesModel, processRowUpdate }}
-          slots={{
-            noRowsOverlay: () => (
-              <Box sx={styles.emptyContent}>No constraints added</Box>
-            ),
-          }}
-          editMode="row"
-          hideFooter
-          checkboxSelection
-          disableColumnMenu
-          disableRowSelectionOnClick
-          onRowEditStop={handleRowEditStop}
-          onCellDoubleClick={handleCellDoubleClick}
-          onRowSelectionModelChange={handleRowSelectionCheckboxChange}
-        />
-        <Button
-          fullWidth
-          size="medium"
-          sx={styles.addBtn}
-          startIcon={<MdAddCircleOutline />}
-          onClick={handleAddRow}
-        >
-          Add a Constraint
-        </Button>
-      </Paper>
-
-      <Stack mt={1} spacing={1} direction="row" justifyContent="end">
-        <Button
-          disabled={!canDiscardOrSaveAll}
-          color="error"
-          variant="contained"
-          startIcon={<MdRestore />}
-          onClick={restoreFilters}
-        >
-          Discard Changes
-        </Button>
-        <Button
-          disabled={!canDiscardOrSaveAll}
-          color="primary"
-          variant="contained"
-          startIcon={<MdCheck />}
-          onClick={handleClickSaveAll}
-        >
-          Save Constraints
-        </Button>
-      </Stack>
-    </>
-  )
 }
 
 // NOTE: This is an unefficient quicker solution
@@ -933,6 +544,22 @@ const FilterModal = ({
   onClose,
 }) => {
   const [filterTab, setFilterTab] = useState('stats')
+
+  const statNames = useSelector(selectGroupedOutputNames)
+
+  const sourceValueOpts = useMemo(
+    () =>
+      R.pipe(
+        R.mapObjIndexed((v, k) => ({ label: k, value: v[1] })),
+        R.values
+      )(statNames),
+    [statNames]
+  )
+  const sourceValueTypes = useMemo(
+    () => R.repeat('number')(sourceValueOpts.length),
+    [sourceValueOpts.length]
+  )
+
   const handleChangeTab = useCallback(() => {
     setFilterTab(filterTab === 'stats' ? 'groups' : 'stats')
   }, [filterTab])
@@ -954,8 +581,10 @@ const FilterModal = ({
         />
       </Tabs>
       {filterTab === 'stats' ? (
-        <StatsFilter
+        <GridFilter
+          sourceHeaderName="Statistic"
           defaultFilters={statFilters}
+          {...{ sourceValueOpts, sourceValueTypes }}
           onSave={handleSaveFilters(groupingFilters)}
         />
       ) : filterTab === 'groups' ? (
