+++ src/ui/views/common/GridEditMultiSelectCell.js
@@ -1,0 +1,8 @@
+import { GridEditSingleSelectCell } from '@mui/x-data-grid'
+
+// TODO: Implement this component to be used in `GridFilter`
+const GridEditMultiSelectCell = (props) => (
+  <GridEditSingleSelectCell {...props} />
+)
+
+export default GridEditMultiSelectCell
