+++ src/ui/views/common/GridFilter.js
@@ -1,0 +1,509 @@
+import { Box, Button, Paper, Stack } from '@mui/material'
+import {
+  DataGrid,
+  GRID_CHECKBOX_SELECTION_COL_DEF,
+  GridActionsCellItem,
+  GridBooleanCell,
+  GridEditBooleanCell,
+  GridEditDateCell,
+  GridEditInputCell,
+  GridEditSingleSelectCell,
+  GridRowEditStopReasons,
+  GridRowModes,
+  useGridApiRef,
+} from '@mui/x-data-grid'
+import * as R from 'ramda'
+import { useCallback, useEffect, useMemo, useState } from 'react'
+import {
+  MdAddCircleOutline,
+  MdCheck,
+  MdDelete,
+  MdEdit,
+  MdRestore,
+  MdSave,
+} from 'react-icons/md'
+import { useSelector } from 'react-redux'
+
+import GridEditMultiSelectCell from './GridEditMultiSelectCell'
+
+import { selectNumberFormat } from '../../../data/selectors'
+
+import { OverflowText } from '../../compound'
+
+import { NumberFormat, mapIndexed, renameKeys } from '../../../utils'
+
+const styles = {
+  addBtn: {
+    justifyContent: 'start',
+    pl: 2,
+    py: 2,
+  },
+  content: {
+    display: 'flex',
+    flexDirection: 'column',
+    height: '100%',
+    my: 2,
+    px: 2,
+    textAlign: 'center',
+    overflow: 'auto',
+  },
+  emptyContent: {
+    display: 'flex',
+    flexDirection: 'column',
+    alignItems: 'center',
+    justifyContent: 'center',
+    bgcolor: 'rgba(18, 18, 18, 0.38)',
+    height: '100%',
+  },
+}
+
+const RELATION_COL_BOOL_VALUE_OPTS = [
+  {
+    value: 'eq',
+    label: 'is',
+  },
+]
+const RELATION_COL_STR_VALUE_OPTS = [
+  {
+    value: 'eq',
+    label: '=',
+  },
+]
+const RELATION_COL_NUM_VALUE_OPTS = [
+  {
+    value: 'lt',
+    label: '<',
+  },
+  {
+    value: 'lte',
+    label: '<=',
+  },
+  {
+    value: 'eq',
+    label: '=',
+  },
+  {
+    value: 'gt',
+    label: '>',
+  },
+  {
+    value: 'gte',
+    label: '>=',
+  },
+]
+const RELATION_COL_SINGLE_SELECT_VALUE_OPTS = [
+  {
+    value: 'eq',
+    label: 'is',
+  },
+]
+const RELATION_COL_MULTI_SELECT_VALUE_OPTS = [
+  {
+    value: 'any',
+    label: 'is any of',
+  },
+]
+const RELATION_COL_DATE_VALUE_OPTS = [
+  {
+    value: 'eq',
+    label: 'is',
+  },
+]
+
+const getRelationValueOptsByType = R.cond([
+  [R.equals('string'), R.always(RELATION_COL_STR_VALUE_OPTS)],
+  [R.equals('number'), R.always(RELATION_COL_NUM_VALUE_OPTS)],
+  [R.equals('boolean'), R.always(RELATION_COL_BOOL_VALUE_OPTS)],
+  [R.equals('singleSelect'), R.always(RELATION_COL_SINGLE_SELECT_VALUE_OPTS)],
+  [R.equals('multiSelect'), R.always(RELATION_COL_MULTI_SELECT_VALUE_OPTS)],
+  [R.equals('date'), R.always(RELATION_COL_DATE_VALUE_OPTS)],
+])
+
+const getCellComponentByType = R.cond([
+  [R.equals('string'), R.always(GridEditInputCell)],
+  [R.equals('number'), R.always(GridEditInputCell)],
+  [R.equals('boolean'), R.always(GridEditBooleanCell)],
+  [R.equals('singleSelect'), R.always(GridEditSingleSelectCell)],
+  [R.equals('multiSelect'), R.always(GridEditMultiSelectCell)],
+  [R.equals('date'), R.always(GridEditDateCell)],
+  [R.T, R.always(GridEditInputCell)],
+])
+
+const GridFilter = ({
+  defaultFilters,
+  sourceHeaderName = 'Source',
+  sourceValueOpts,
+  sourceValueTypes,
+  onSave,
+}) => {
+  const [filters, setFilters] = useState(defaultFilters)
+  const [idCount, setIdCount] = useState(0)
+  const [rows, setRows] = useState([])
+  const [initialRows, setInitialRows] = useState([])
+  const [rowModesModel, setRowModesModel] = useState({})
+  const [canSaveRow, setCanSaveRow] = useState({})
+
+  const apiRef = useGridApiRef()
+  const numberFormat = useSelector(selectNumberFormat)
+
+  const sourceTypesByValue = useMemo(
+    () =>
+      R.pipe(
+        R.pluck('value'),
+        R.flip(R.zipObj)(sourceValueTypes)
+      )(sourceValueOpts),
+    [sourceValueOpts, sourceValueTypes]
+  )
+
+  const isApiRefValid = !R.either(R.isNil, R.isEmpty)(apiRef.current)
+
+  const restoreFilters = useCallback(() => {
+    const filterCriteria = R.filter(
+      R.pipe(R.propOr('stat', 'format'), R.equals('stat'))
+    )(filters)
+    const initRows = mapIndexed(
+      R.pipe(
+        R.flip(R.assoc('id')),
+        // Set up `logic` for first row in case it is `undefined`
+        R.when(R.propEq(0, 'id'), R.assoc('logic', 'and')),
+        // Drop truthy `active` values
+        R.when(R.propEq(true, 'active'), R.dissoc('active')),
+        renameKeys({ option: 'relation', prop: 'source' })
+      )
+    )(filterCriteria)
+
+    setInitialRows(initRows)
+    setRows(initRows)
+    setIdCount(initRows.length)
+  }, [filters])
+
+  useEffect(() => {
+    restoreFilters()
+  }, [restoreFilters])
+
+  useEffect(() => {
+    if (!isApiRefValid) return
+
+    const newSelectedRowIds = R.pipe(
+      R.filter(R.propOr(true, 'active')),
+      R.pluck('id')
+    )(initialRows)
+    apiRef.current.setRowSelectionModel(newSelectedRowIds)
+  }, [apiRef, initialRows, isApiRefValid])
+
+  const preProcessEditCellProps = useCallback(
+    ({ id, props, hasChanged, otherFieldsProps }) => {
+      const isValueInvalid = R.propSatisfies(
+        R.either(R.isNil, R.isEmpty),
+        'value'
+      )
+      const newCellProps = R.pipe(
+        R.ifElse(isValueInvalid, R.assoc('error', true), R.dissoc('error'))
+      )(props)
+
+      if (hasChanged) {
+        const hasError =
+          newCellProps.error ||
+          R.pipe(
+            // We ignore the `logic` field in the first row
+            R.when(R.always(id === rows[0].id), R.dissoc('logic')),
+            R.values,
+            R.any(isValueInvalid)
+          )(otherFieldsProps)
+
+        setCanSaveRow(R.assoc(id, !hasError))
+      }
+      return newCellProps
+    },
+    [rows]
+  )
+
+  const handleAddRow = useCallback(() => {
+    setRows(
+      R.append({
+        isNew: true,
+        id: idCount,
+        source: '',
+        relation: '',
+        logic: 'and',
+      })
+    )
+    setRowModesModel(
+      R.assoc(idCount, {
+        mode: GridRowModes.Edit,
+        fieldToFocus: rows.length < 1 ? 'source' : 'logic',
+      })
+    )
+    apiRef.current.selectRow(idCount)
+    setIdCount(idCount + 1)
+  }, [apiRef, idCount, rows.length])
+
+  const deleteRow = useCallback((id) => {
+    setRows(R.reject(R.propEq(id, 'id')))
+  }, [])
+  const handleDeleteRow = useCallback(
+    (id) => () => {
+      deleteRow(id)
+    },
+    [deleteRow]
+  )
+
+  const handleClickDiscard = useCallback(
+    (id) => () => {
+      const row = apiRef.current.getRow(id)
+      setRowModesModel(
+        R.assoc(id, { mode: GridRowModes.View, ignoreModifications: true })
+      )
+      if (row.isNew) deleteRow(id)
+    },
+    [apiRef, deleteRow]
+  )
+  const handleClickEdit = useCallback(
+    (id) => () => {
+      setRowModesModel(R.assoc(id, { mode: GridRowModes.Edit }))
+    },
+    []
+  )
+  const handleClickSave = useCallback(
+    (id) => () => {
+      setRowModesModel(R.assoc(id, { mode: GridRowModes.View }))
+      setCanSaveRow(R.dissoc(id))
+    },
+    []
+  )
+
+  const processRowUpdate = (newRow) => {
+    const updatedRow = R.dissoc('isNew')(newRow)
+    setRows(R.map(R.when(R.propEq(newRow.id, 'id'), R.always(updatedRow))))
+    return updatedRow
+  }
+
+  const handleRowEditStop = ({ reason }, event) => {
+    if (reason === GridRowEditStopReasons.rowFocusOut) {
+      event.defaultMuiPrevented = true
+    }
+  }
+
+  const updateActiveStateForSelectedRow = useCallback(
+    (row) => {
+      const selectedRowIdsSet = new Set(apiRef.current.getSelectedRows().keys())
+      return R.ifElse(
+        (row) => selectedRowIdsSet.has(row.id),
+        R.dissoc('active'),
+        R.assoc('active', false)
+      )(row)
+    },
+    [apiRef]
+  )
+
+  const canDiscardOrSaveAll = useMemo(() => {
+    const unsavedChanges = !R.equals(rows)(initialRows)
+    const isEditing = R.pipe(
+      R.values,
+      R.any(R.propEq(GridRowModes.Edit, 'mode'))
+    )(rowModesModel)
+    return !isEditing && unsavedChanges
+  }, [initialRows, rowModesModel, rows])
+
+  const handleClickSaveAll = useCallback(() => {
+    const newFilters = R.map(
+      R.pipe(
+        updateActiveStateForSelectedRow,
+        R.dissoc('id'),
+        renameKeys({ relation: 'option', source: 'prop' })
+      )
+    )(rows)
+    setFilters(newFilters)
+    onSave(newFilters)
+  }, [onSave, rows, updateActiveStateForSelectedRow])
+
+  const handleRowSelectionCheckboxChange = useCallback(() => {
+    setRows(R.map(updateActiveStateForSelectedRow))
+  }, [updateActiveStateForSelectedRow])
+
+  const handleCellDoubleClick = (params, event) => {
+    event.defaultMuiPrevented = true
+  }
+
+  const columns = useMemo(
+    () => [
+      {
+        field: 'logic',
+        headerName: 'Logic',
+        align: 'center',
+        width: 90,
+        editable: true,
+        sortable: false,
+        type: 'singleSelect',
+        valueOptions: ['and'], // FIXME: ['or', 'and'],
+        getOptionLabel: (option) => option.toUpperCase(),
+        cellClassName: ({ id }) => (id > rows[0].id ? '' : 'hidden'),
+        preProcessEditCellProps,
+      },
+      GRID_CHECKBOX_SELECTION_COL_DEF,
+      {
+        field: 'source',
+        headerName: sourceHeaderName,
+        type: 'singleSelect',
+        width: 180,
+        editable: true,
+        sortable: false,
+        valueOptions: sourceValueOpts,
+        renderCell: ({ formattedValue }) => (
+          <OverflowText text={formattedValue} />
+        ),
+        preProcessEditCellProps,
+      },
+      {
+        field: 'relation',
+        headerName: 'Relation',
+        headerAlign: 'center',
+        align: 'center',
+        width: 90,
+        type: 'singleSelect',
+        editable: true,
+        sortable: false,
+        valueOptions: ({ row, id }) => {
+          const valueAlt = R.path([id, 'source'])(rows)
+          const value = R.propOr(valueAlt, 'source')(row)
+          const valueType =
+            value && value !== '' ? sourceTypesByValue[value] : 'number'
+          return getRelationValueOptsByType(valueType)
+        },
+        preProcessEditCellProps,
+      },
+      {
+        field: 'value',
+        headerName: 'Value',
+        width: 140,
+        type: 'number',
+        editable: true,
+        sortable: false,
+        valueFormatter: ({ value }) =>
+          typeof value === 'number'
+            ? NumberFormat.format(value, numberFormat)
+            : null,
+        renderCell: (params) => {
+          const valueType = sourceTypesByValue[params.row.source]
+          return valueType === 'boolean' ? (
+            <GridBooleanCell {...params} />
+          ) : (
+            <OverflowText text={`${params.value}`} />
+          )
+        },
+        renderEditCell: (params) => {
+          const valueType = sourceTypesByValue[params.row.source]
+          const Component = getCellComponentByType(valueType)
+          return <Component {...params} />
+        },
+        preProcessEditCellProps,
+      },
+      {
+        field: 'actions',
+        headerName: 'Actions',
+        headerAlign: 'right',
+        align: 'right',
+        width: 100,
+        type: 'actions',
+        sortable: false,
+        getActions: ({ id }) =>
+          R.path([id, 'mode'])(rowModesModel) === GridRowModes.Edit
+            ? [
+                <GridActionsCellItem
+                  disabled={!canSaveRow[id]}
+                  icon={<MdSave size="20px" />}
+                  label="Save"
+                  onClick={handleClickSave(id)}
+                />,
+                <GridActionsCellItem
+                  icon={<MdRestore size="20px" />}
+                  label="Discard"
+                  onClick={handleClickDiscard(id)}
+                />,
+              ]
+            : [
+                <GridActionsCellItem
+                  icon={<MdEdit size="20px" />}
+                  label="Edit"
+                  onClick={handleClickEdit(id)}
+                />,
+                <GridActionsCellItem
+                  icon={<MdDelete size="20px" />}
+                  label="Delete"
+                  onClick={handleDeleteRow(id)}
+                />,
+              ],
+        preProcessEditCellProps,
+      },
+    ],
+    [
+      canSaveRow,
+      handleClickDiscard,
+      handleClickEdit,
+      handleClickSave,
+      handleDeleteRow,
+      numberFormat,
+      preProcessEditCellProps,
+      rowModesModel,
+      rows,
+      sourceHeaderName,
+      sourceValueOpts,
+      sourceTypesByValue,
+    ]
+  )
+
+  return (
+    <>
+      <Paper sx={styles.content}>
+        <DataGrid
+          {...{ apiRef, columns, rows, rowModesModel, processRowUpdate }}
+          slots={{
+            noRowsOverlay: () => (
+              <Box sx={styles.emptyContent}>No constraints added</Box>
+            ),
+          }}
+          editMode="row"
+          hideFooter
+          checkboxSelection
+          disableColumnMenu
+          disableRowSelectionOnClick
+          onRowEditStop={handleRowEditStop}
+          onCellDoubleClick={handleCellDoubleClick}
+          onRowSelectionModelChange={handleRowSelectionCheckboxChange}
+        />
+        <Button
+          fullWidth
+          size="medium"
+          sx={styles.addBtn}
+          startIcon={<MdAddCircleOutline />}
+          onClick={handleAddRow}
+        >
+          Add a Constraint
+        </Button>
+      </Paper>
+
+      <Stack mt={1} spacing={1} direction="row" justifyContent="end">
+        <Button
+          disabled={!canDiscardOrSaveAll}
+          color="error"
+          variant="contained"
+          startIcon={<MdRestore />}
+          onClick={restoreFilters}
+        >
+          Discard Changes
+        </Button>
+        <Button
+          disabled={!canDiscardOrSaveAll}
+          color="primary"
+          variant="contained"
+          startIcon={<MdCheck />}
+          onClick={handleClickSaveAll}
+        >
+          Save Constraints
+        </Button>
+      </Stack>
+    </>
+  )
+}
+
+export default GridFilter
